feat(scraper): add `pino` logs and `user-agents` spoofing
This commit adds `pino` logging and uses the `user-agents` package to
attempt to trick PerimeterX into thinking we're a real user by setting a
realistic `User-Agent` request header. This attempt was to no avail and
I'm still getting Captchas, so I'm going to be trying to use the
Puppeteer stealth plugin next. If that fails, I'll resort to copying my
JQuery code into my browser console directly and simply downloading the
resulting JSON. That process would be feasible for my goals right now.