feat(scraper): use puppeteer to enable interactions
This commit installs and uses Puppeteer to enable JS enabled
interactions that are unavailable by simply using Cheerio and the
Node.js fetch API. This approach is quite a bit slower but should be
more robust for complex scraping interactions.

Ex: Clicking the "LOAD MORE" button on Isabel Marant's website.