feat(prisma): add product colors and pricing
This commit creates a new "Color" model to encapsulate the color labels
that designers assign to their products. The "Product" and "Price"
models now have an array of "Color" models each (each product can have
many colors it was originally made in and each price can be for one or
more colorways).