feat(scraper): finish basic marant scraping
This commit implements a functioning Isabel Marant scraping script. This
script, however, is probably not the most performant as it can be; I
suspect that a lot of the time each iteration takes is loading the page
which I can get rid of by simply hitting the "Clear filters" button each
time instead of spinning up a new page.