ci(workflows/deploy): use `prod` as production branch
This commit removes the `dev` branch and instead marks `main` as the
development branch while `prod` is the production branch. This change is
purely aesthetic and was made primarily just to maintain consistency
with what I'm having to do at work (so my workflows remains the same and
I don't have to context switch as much).