fix(components): remove `autoFocus` prop for a11y
This commit removes the `autoFocus` property which apparently is bad for
a11y reasons (and thus was causing my ESLint to fail). In the future,
I'll have to implement a more robust auto-focusing system as I
definitely still want users to be able to interact with the app in its
entirety without having to use a mouse. Keyboard shortcuts for the win!