refactor(data/marant): organize directories by date
This commit creates separate directories for every date that I've
scraped data from Isabel Marant. I believe this approach is better than
simply overwriting the data tracked in git as currently we're replacing
the previous JSON files and not appending to it.

Ideally, our scraper scripts will simply modify the existing JSON files
(appending to data and updating the "last seen at" fields for duplicate
products). Our git history should essentially track changes to the
websites we scrape. To do this, we'll need to update our JSON schemas
and our scraping scripts: I'll do that in a later commit.