refactor(data): make data publicly available on site
This commit moves the data directory to the `public` directory so that
it is publicly hosted on our website for convenient downloads. This
change also makes sense if we're going to be downloading product images
alongside the JSON data at scrape time and want to be able to use those
image URLs at runtime.