feat(scraper/marant): save product images locally
This commit adds a new `saveImages` function to our Isabel Marant
scraper script to download product images locally.

This resolves an issue where runtime fetches to the Isabel Marant hosted
images would hang and thus image optimization didn't work. This also
ensures that we are not dependent on 3rd parties for assets.