ci: add `semantic-release` bot for releases
This patch adds a new release workflow which monitors commits to main.
When commits are made, it uses semantic-release [[1]] to analyze the git
commits (assuming we use conventional commits) since the last release,
generate a CHANGELOG, and a version number based on the commit subjects.

It will then push a new commit directly to `main` containing the updates
to the CHANGELOG as `chore(release): {new version}`. It will then tag
this commit with the new version number and push the tag.

Note that be default, it will make the first release v1.0.0. This is
true even though `v8` tag exists.

For the dry-run, we have to checkout the branch awkwardly due to this
issue: [[2]]

[1]: https://semantic-release.gitbook.io/semantic-release/
[2]: https://github.com/semantic-release/semantic-release/issues/1890#issuecomment-974512960