perf(filters): reuse already mounted field items
This patch updates the filters add button component to reuse the already
mounted field items when a specific field is selected and to simply hide
the other irrelevant field items (when a field is selected). This change
removes a UX flash where items are being fetched in ~50ms after a
specific field is selected.