feat(cypress): add aritzia scraping script
This patch creates a scraping script to get product information from
Aritzia using Cypress to drive a headless browser. This approach works
splendidly and is fantastic for quick iteration cycles but introduces
quite a bit of unnecessary overhead to a simple HTML scrape; Aritzia
does not have bot protections (no PerimeterX to be found here) and does
not require JS to scrape (as they populate everything in convenient
`data-` attributes).