refactor(scraper): name directories by language
This patch refactors the scraper directory to name the child directories
by the language that they are written in. This is better than naming
them by `browser` and `script` as I want to reuse the Node.js TS
configuration for random Node.js scripts that do not necessarily have
anything to do with headless browser driving.