ci(workflows/deploy): remove deploy and just build
This patch updates the deploy workflow to build the docker image in
GitHub Actions instead of on a Fly.io builder machine and uploads the
resulting image to GHCR instead of the Fly.io registry.

This change also removes the Fly.io deploy step as I'll be deploying to
either a very small Docker hosting service (e.g. DigitalOcean droplet)
or Vercel without configuration.