ci(workflows/deploy): remove `permissions` key
Previously, the `permissions` key was preventing writing to GHCR and I
wanted to stay consistent with the other GitHub Actions in this
repository so I simply removed the key altogether. I may add it back in
the future (but if I do, it will be to every workflow in this project).