feat: add serif fonts for decorative headers
This patch adds the "Bodoni Moda" serif font for use in decorative
headers (e.g. in the show page). This was inspired by Vogue and other
fashion publications who use serif fonts for their logos and large
headers. I chose "Bodoni Moda" as it was freely available on Google
Fonts, looked similar to the "FBDidotS" font that Vogue uses in their
headers, and was available as a variable font.