fix(routes/show): move about to the left
This patch switches the "about" panel (with the scores and collection
title) to the left side as most users will be reading left-to-right.
Eventually, I should decide which direction to put these based on the
LTR / RTL orientation of the user's i18n. But that'll only be added once
I have real i18n setup... which will likely be never. :)