feat(routes/shows): set look image aspect ratio
This patch sets a fixed look image aspect ratio (9 / 16) to avoid layout
shift and to ensure that the "shows" grid has consistent image sizes. I
may decide to revert this change for the "looks" grid in the actual show
page (as designers may have an artistic choice in framing their look
photos), but for now, this works elegantly.

This change was inspired by the fixed ratio images used in the Vogue
Runway list of shows [[1]].

[1]: https://www.vogue.com/fashion-shows