feat(prisma): add critic review fields
This patch adds the title, subtitle, summary, and score fields to the
`Review` object. Similar to how Rotten Tomatoes assigns a sentiment
(i.e. rotten/fresh) to each critic review, we should do the same. We
should also save the title and subtitle from the review in our database.
And we should include a short summary of the review that is shown to
users until they expand the review to "read more" (similar to Rotten
Tomatoes).

Closes: NC-629