fix(routes): verify there are no existing users
This patch updates the sign up flow to ensure that we don't let users
create accounts that already exist or have unique constraint conflicts.
We can't catch everything, however, and still need to work on adding
better error surfacing for non-technical users.

This patch also ensures that the submit buttons are disabled while
submitting the form. I also tried disabling the inputs but that was
messing with the server-side error surfacing in the front-end.