feat: add `show.url` and `retailer.company`
This patch adds a unique URL field to the `Show` model. This URL should
indicate the link to the show page on the brand's website. This page
should be where the `show.description` comes from.

This patch also updates the `Retailer` model to include a `company`
field. This field should be marked as required once the existing
database has been migrated over (i.e. once I've added `company`
relations to all of the existing retailers).

This patch updates the corresponding scraper scripts to specify the show
URL and collection link(s) (if applicable) when importing shows.