feat(routes/show): add "where to buy" section
This patch adds a "where to buy" section to the show page. Similar to
how Rotten Tomatoes has a "where to watch" section on their movie
overview page [[1]], this "where to buy" section gives users one-click
links to shop the looks and products presented in the show.

The "where to buy" section only includes direct links to the specific
collection(s) presented in the show (e.g. for Hermès Spring-Summer 2023,
that link may be [[2]]).

If a direct link to the collection presented in the show does not exist
(e.g. often, Nordstrom and Neiman Marcus and similar retailers will not
maintain a dedicated page per brand collection), we simply show an
`<Empty>` component that has a link to the brand website so users can go
find the products themselves.

Eventually, that behavior will be enhanced once we have the whole:

  Show > Look > Product > Variant > Price (with link to purchase)

...data model hierarchy properly fleshed out in our database. For now,
this works as the next-best option without requiring extensive scraping
to add a show to the database.

[1]: https://www.rottentomatoes.com/m/oppenheimer_2023
[2]: https://www.hermes.com/us/en/category/men/ready-wear/spring-summer-collection

Closes: NC-624