fix(routes/show): style "where to buy" alt text
This patch fixes an issue where the "where to buy" button images were
preventing users from scrolling the show page when the image failed to
load and the `alt` text was shown. Now, if the image fails to load or if
there isn't a valid `avatar` to load, the button simply looks like a
text outlined button.