feat: calculate show scores based on review scores
This patch updates the show page to calculate the overall show scores
based on the review score counts that are positive and negative. This
change better mimics what Rotten Tomatoes does and will help prevent
outliers from skewing averages, etc.

Note that this change made me update the `/shows` page to no longer show
a preview of the show scores as I did not want to dynamically calculate
them for each show on that page as well (that would've required quite a
few database selects via Prisma or some complex SQL I do not want to
write right now).

Closes: NC-641