feat(routes/show): add consumer review sections
This patch adds the "Rate and review" section to the show page where
users can submit their own consumer reviews.

This patch also adds the "Consumer reviews" section which contains a
very basic list of the consumer reviews currently in database.

This patch includes a couple other small drive-by fixes and linking
improvements (e.g. we now associate an ID with each section on the show
page, making it easier to provide `#` based links to specific parts of
the increasingly long page).

Closes: NC-625