fix(show/where-to-buy): invert logos in dark mode
This patch updates the "where to buy" buttons to invert their brand or
retailer logos when the user switches to dark mode. Often, these logo
images will simply be PNGs or SVGs with black text on a clear
background. Inverting them gives us a nice logo for dark mode: white
text on a clear background.

Fixes: a909447a4583 ("feat(routes/show): add "where to buy" section")