fix(routes): mark login and sign up fields required
This does nothing to the form validation functionality (we still will
use `@conform-to` client and server-side) but it ensures that the
`aria-required` attributes are set correctly. When not using JS
client-side, this results in a slightly better experience as the user's
browser will do some client-side validation for us before our server
performs validation.