feat(routes/show): add star-based score input
This patch replaces the simple `<Input type='number' />` previously used
to submit scores with a star-based selector that uses a radio group and
`peer-` modifiers under the hood for CSS-only functionality.

Note that this selector will not work without JS, as the Radix radio
group component does not work without JS either. But the CSS-only
styling should be incredibly performant.