fix(routes/show): links to sections show headers
This patch updates the show page styling so that an anchor-based link
(`#`) to a specific section on the show page will include that show's
header. Previously, the show's header was being hidden by the top app
bar. Now, I moved the gap between sections to the section padding
itself, which makes the section headings appear when linking directly to
them.