fix: users can only submit one review per show
This patch adds a unique constraint on the `showId` and `authorId`
fields on the `Review` table in Postgres. This patch then updates the
"rate and review" section to show the user's existing review if it
exists (along with a submit button that reads "Edit review").

Users can only submit a single review per show.

Closes: NC-643