feat(show/designers): collapse description by default
This patch updates the "designers" section of the show page to collapse
the description by default and only show the entire thing if the user
clicks on the "show more" button.

Note that this implementation is WIP and does not have the best a11y,
but it works perfectly for the styles that I want. Eventually, I may
want to try using the `<Collapsible>` from Radix Primitives, but this
works for now.