feat: add `nprogress` page transition indicator
This patch adds the nprogress page transition loading indicator. This is
a temporary fix that should be addressed by simply making page
transitions instantaneous. To do that, I will want to implement some
type of caching or ISR strategy, or simply move back to using Fly.io on
the edge (so my database is right next to my Node.js server).

Closes: NC-642