fix: ensure look images fully cover wrappers
This patch updates the scores header to ensure that the preview look
image properly covers the entire section to fill. Previously, this issue
was not so readily apparent until I started importing all the shows from
Vogue. Not every show on Vogue has the same aspect ratio of look image.
This ensures that they appear to, at least, on my website.