feat(prisma): add review "written at" date field
This patch adds a new `date` field to the `Review` object. This date
field will be `NULL` for consumer reviews (as the written at date for
consumer reviews will simply be createdAt/updatedAt). For critic
reviews, this is a separate field as most of the time, the date when the
critic review was originally written will be different from the date I
imported it.