fix: make the show description field optional
The show description will be NULL by default and only populated if the
designer or brand has provided a description. Otherwise, it will be left
empty. I want to separate reviews from the show description. Previously,
I was setting the show description to the Vogue review content during
import. That resulted in large duplicate blocks of text in the UI.