fix(routes/show): hide empty "about" sections
This patch hides the empty sections in the left side of the show page.
Rotten Tomatoes, when there is no critic or consumer summary, will
simply hide that section on the movie page. This change implements
similar behavior for almost every section of the "about" left side of
the show page.

Note that I did not delete the `<Empty>` state as I may want to
re-introduce some of them at some point (e.g. the "where to buy" empty
state can sometimes actually be helpful, pointing users to the brand's
website where they may be able to find the collection's products).