refactor(show/consumer-reviews): break into component
This patch creates a new `<ConsumerReview>` component that can be reused
(e.g. in the profile page to provide users with a preview of their
profile edits) across the application.

This patch also re-exports the `<TimeAgo>` component from our own
component file (to ensure that the `en_short` locale is included
wherever we use `timeago-react`).