feat(show/og): add show specific social images
This patch adds an OG image generation route for shows. It will generate
an image with the show name and a preview of the first three looks in
the show. This is a better experience than Vogue, in my opinion, as
Vogue simply crops the first look image in an awkward manner (because
the social media image aspect ratio is not very compatible with a look
image ratio).