perf(routes): lazy load show and look images
This patch updates the show and shows list routes to lazy load the look
and show images respectively. This patch also adds a responsive `srcset`
to the shows list page. This patch also adds a somewhat redundant
`preload='auto'` attribute to the show runway video (redundant because
we're already telling the browser to preload the video by setting auto
play to true).

Closes: NC-675
Closes: NC-676