feat(routes): add user friendly show paths
This patch replaces the `$showId` path with a more user-friendly URL
path for shows that includes the show season year, name, the show sex
(i.e. "man" or "woman"), and the show brand slug. This mimics the
behavior of Vogue, but with more nesting (which I think is better).

Note that there is still some discussion to be had on what the order of
that nesting should be (e.g. are "man" and "woman" top-level categories,
even higher up the hierarchy than the season year?) but this current
setup makes sense, as it is how you would read it: "2023's fall menswear
collection by Acne Studios".

I also have made the choice that collections for women will omit the sex
identifier from the URL path. Most collections are for women. This
choice mimics what Vogue and other publishers do. However, once I have
successfully combined the duplicate menswear collections with their
female counterparts (see NC-660), I should revert this change and
include "woman" in the URL path and omit the sex field when it is
unisex.

Ref: https://developers.google.com/search/docs/fundamentals/seo-starter-guide?visit_id=638258623280664328-3236374080&rd=1#simpleurlsconveyinfo
Ref: https://developers.google.com/search/docs/fundamentals/seo-starter-guide?visit_id=638258623280664328-3236374080&rd=1#navisimportant

Closes: NC-673