fix(routes): replace "login" with "log in"
This patch replaces the noun "Login" with the verb phrase "Log in" to
better represent what we're doing. We are not referring to your "Login"
(noun; meaning username + password) but rather that you _can_ "log in"
to the app by clicking a button or link to the log in page.

Ref: https://grammarist.com/spelling/log-in-login/
Ref: https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Attributes/aria-label