test(e2e/smoke): add basic login and profile tests
This patch adds a basic log in and log out smoke test. I will add show
review tests next. Note that this interacts with the front-end as if I
was a user and does not make any assertions about network requests. If
the app gets slow enough to require that, I should simply speed it up.

Ref: https://testing-library.com/docs/guiding-principles
Ref: https://twitter.com/kentcdodds/status/977018512689455106