feat(scripts/vogue): import all vogue show data
This patch updates the Vogue Scrapy spider to start on the seasons index
page (that lists all of the seasons on Vogue) and then branch off from
there. This should ensure that we have all the shows in our database
that Vogue has reviews on (at least, that Vogue has posted on their
public, production website).

Next up: WWD... :)

Closes: NC-644