fix(prisma): add `location` to show unique constraint
This patch adds the "location" field to the show unique constraint. This
is required as the same brand _can_ have two different shows (with two
different collections) in the same season but in different locations.

For example, Maticevski had an Australia Spring 2015 RTW collection that
was distinct from (and included different looks than) their Paris Spring
2015 RTW collection (see links included below).

Ref: https://www.vogue.com/fashion-shows/spring-2015-ready-to-wear/maticevski
Ref: https://www.vogue.com/fashion-shows/australia-spring-2015/maticevski