fix(shows): sort by brand name and then by year
This patch updates the shows list page to sort by the show brand name
instead of the show name itself. This allows me to then sort by the show
year (which previously was conflicting with the name filters as the show
names include the show year).

Eventually, I should probably just get rid of the show name field as it
can introduce confusion. It is useful for ensuring duplicate shows are
not created while importing, however (e.g. I can `connectOrCreate` by
the show name which is deterministically set, instead of having to link
by ID or the combination of a bunch of fields).