fix(scripts/vogue): filter out duplicate shows
This patch adds some filtering to the Vogue Node.js save script to
attempt to filter out duplicate shows before attempting to create those
records in database. Note that this does not work 100% of the time, as
some shows do not _exactly_ match but rather _mostly_ match. In those
cases, I could use something like a levenshtein distance to filter them
out programmatically, but doing so would not be performant.

Instead, I'll just log those Postgres errors and inspect them to see if
I need to correct something in my schema or script or if the failure was
an expected duplicate and can be safely ignored.