perf(shows): virtualize and add infinite scrolling
This patch virtualizes the shows page grid and adds infinite scrolling
so that only the necessary subset of data is loaded and rendered at any
given time. This change was necessary as I now have 20,857 fashion
shows, which should not be rendered all at once.

This implementation was heavily inspired by the masonry example on the
Tanstack documentation and the infinite scrolling Remix example.

Note that this is a purely FE performance improvement; the PostgreSQL
queries will not be _fully_ optimized as I was forced to use offset
based pagination instead of cursor based pagination (due to the custom
ordering the shows page requires).

This patch also ensures that the user's scroll position on the shows
page is restored when navigating from anywhere (including breadcrumbs).

Ref: https://tanstack.com/virtual/v3/docs/examples/react/variable
Ref: https://github.com/remix-run/examples/blob/main/infinite-scrolling/app/routes/offset/advanced.tsx
Ref: https://www.prisma.io/docs/concepts/components/prisma-client/pagination#-cons-of-offset-pagination

Closes: NC-691
Closes: NC-659