fix(scripts/wwd): do not include location in name
This patch updates the WWD import script to omit the location from the
show name when matching shows with the existing ones in Postgres. Vogue
does not include show locations for most of their shows. The only shows
that _actually_ caused conflicts were the Hyke Tokyo shows. I updated
the script accordingly.

Fixes: 5d3e9b59b8ed ("feat(scripts/wwd): add import script and migrations")

Closes: NC-693