fix(root): remove link to private repository
This patch removes the GitHub Issues creation link from the error page.
The repository is currently private, and thus this link will simply 404.
The inclusion of this link also doxes myself... no need to do so
unnecessarily. The only link back to me now could _perhaps_ be a couple
comments in code, but those _should_ be stripped during the build
process, so I should be good.