fix(utils): update `BASE_URL` to `dolce.so`
This patch updates the `BASE_URL` configuration to correctly point to
the new domain name for "dolce". This is used primarily for the URL
fields in the external rich metadata that is used by Google and other
search engines for rendering rich search results (e.g. for the show
page).