feat(show/looks): add "save" icon button toggle
This patch adds a "save" icon button to each look in a show that adds
the look to the user's "DEFAULT" set (or creates that set if it does not
yet exist). Eventually, I will add a drop-down menu where users can
select which collection they would like to add the selected look to.

Closes: NC-637