feat(routes/show): add convenience redirect route
This patch adds a new `shows/$showId` route that simply redirects to the
actual show page based on the ID. This is useful when writing redirects
when the show ID is easily accessible but I do not want to select all of
the nested attributes that are required to create the full show URL
unnecessarily (i.e. no need to fetch all those nested relations to
construct a link that the user may not even click).

This patch adds a route that will do all that nested relation fetching
and redirect to the correct show page given a valid show ID. The search
parameters and hash of the original URL are preserved correctly.