feat(routes/user): add public user profile page
This patch creates a new public user profile page heavily inspired by
the Instagram website public profile page. This patch also includes a
slight update to the `<Avatar>` component to allow me to change the
placeholder font-size from the root `className`.

This patch also updates the global header icons to include a separate
settings icon button that opens the profile settings page and updates
the user icon to open the current user's public profile page.

This patch includes some useless placeholder "follow" and "message"
buttons that are only shown to unauthenticated users and simply redirect
to the `/join` page. Eventually, these will do something for logged in
users. For now, they're useful to show the possibilities with the
product.

Closes: NC-714