feat: add "sets" multi-select when saving a look
This patch adds a multi-select that opens when a user clicks the "save"
icon button. This multi-select allows users to filter their sets by name
and select which one(s) they would like to save the look to.

To implement this, I've added a few components from the `shadcn-ui`
collection (`Command`, `Dialog`, and `Popover`). I also added an API
endpoint to list the first 100 sets that include the user's query in
their name. Note that I decided that adding full text-based search was
out-of-scope for this particular feature (the set names shouldn't be too
complex... and, most likely, a single user won't have 100+ sets).

Ref: https://ui.shadcn.com/docs/components/combobox
Ref: https://cmdk.paco.me/
Ref: https://github.com/pacocoursey/cmdk#asynchronous-results
Ref: https://www.prisma.io/docs/concepts/components/prisma-client/full-text-search#postgresql

Closes: NC-713