fix(utils): don't check user type in `useUser`
This patch removes the `isUser` check in the `useOptionalUser` hook.
This was causing some bugs when there was a valid `User` object that did
not have an email address and thus did not pass the `isUser` check.

I can rely on TS here. I don't think I need a runtime type guard.