fix(routes): use `redirectTo` when already logged in
This patch updates the join and login pages to respect the `redirectTo`
search parameter (if it exists) when a user is already logged in.

This fixes an issue when a user opens multiple tabs, logs into one, and
then tries to save an item in another tab. Previously, clicking the save
icon would redirect to the index page. Now, it simply refreshes the
current page. I should ideally fix this by adding some sort of tab focus
revalidation so that users don't have to click the save button twice
(once to refresh the page and replace it with the select, and once to
open the now-replaced select).