feat(routes/index): add location popovers to map
This patch adds location popovers that open when you click on a location
on the homepage map. The popover shows the most recent shows in that
location, and the show's statistics (i.e. the number of shows at that
location and the number of unique brands who have had shows there).