fix(routes/index): remove map projection skew
This patch removes some weird rotation from the default map projection
that seemed unnecessary. This was simply copied over from the example
and I'm not sure if it is necessary. Perhaps it resulted in a more
accurate map previously, but I like the way that it looks now better and
accuracy is not a large concern for this particular map application.