feat(routes/shows): show results load time
This patch updates the shows header to be quite a bit smaller and to
mimic the Google search results subheader that shows the number of
results and the load time for those results. To do this, I'm showing the
TTFB web vital metric. There is probably some polish to be done there,
but I think it's somewhat accurate for now and provides a nice sense of
emphasis that I've loaded so many results in such a short time (usually
around 20K results in 0.5s to 2s at most).