fix(routes/show): ensure content is centered
This patch ensures that the content in the show page is properly
centered in the user's viewport. Previously, while it may have appeared
somewhat centered, it was not exactly center.

Ideally, I should have done this with pure CSS manipulation, but doing
so proved difficult (and may be impossible), so I instead simply
calculate out the proper `flexGrow` attributes on window resize.

Fixes: 632608e86bb1 ("fix(routes/show): allow scrolling on padding areas")

Closes: NC-661