fix(show/rate-and-review): replace `action` with `useFetcher`
This patch removes the action function to a `useFetcher` call to a show
specific API endpoint that includes the `showId` in the path parameters.
Previously, there was an issue where the show could not be found as the
path parameters in the action were renamed. We are no longer supporting
non-JS platforms, so this change made more sense than fixing the
original `action`.