fix: `useLayoutEffect` for DOM measurements
This patch replaces the `useEffect` in the show and shows page that is
used for window viewport measurements with a `useLayoutEffect` to avoid
UI flickers after the page has loaded.

There still will be a UI flicker on initial page load (at least, for the
show page). I should ideally address this by implementing a CSS-only
centering solution instead of relying on viewport measurements.

Fixes: 1686312acc0d ("fix(routes/show): ensure content is centered")