fix(button): keep transparent color in dark mode
This patch updates the `<Button>` component to keep the same transparent
color when in dark mode that is used in light mode. This variant is only
ever used on top of already colorful content (e.g. images or videos) and
thus switching the background color in dark mode doesn't make sense.