feat(map): add pagination dots to location popover
This patch adds pagination indicator dots to the location popover that
indicate to the user how many shows or items in the carousel are left.

This change was heavily inspired by the Airbnb rental popover (where you
can click through the images of the home while still in the map view).

Ref: https://www.airbnb.com/