fix(map): manually reposition to avoid conflicts
This patch updates the map location coordinates to be somewhat
inaccurate to avoid conflicts where two locations were directly on top
of one another. This map does not need to be accurate, and I would
rather have inaccurate locations that can be clicked than accurate
locations that overlap with one another.