fix(map): hide scrollbar in location shows carousel
This patch hides the scrollbar in the location shows carousel. Users can
still swipe to scroll (e.g. on mobile or on a magic trackpad) but are
encouraged to use the pagination buttons instead. This also ensures that
the styling looks good across platforms (e.g. windows always shows an
ugly scrollbar).