fix: put automatic revalidation behind an env flag
This patch disables the automatic revalidation on focus and reconnect by
default and adds two environment variables that can be set to re-enable
it. I was experiencing issues with this where the revalidation would
often fail (e.g. due to Vercel timeout issues) and pages would appear to
simply break on re-focus.

Enabling this during development is also not an ideal experience, as I
no longer take advantage of the Remix HDR (I was revalidating all data
every time I switched between my terminal and Firefox).

Fixes: 0cf43042ad0b ("feat: automatically revalidate on focus or reconnect")