perf(shows): reduce default take from 200 to 100
This patch reduces the default take on the shows page from 200 to 100.
Now that I am also getting all of the nested looks for each show, I want
to load less with each request. This number is still adequate and should
be hard to outrun (i.e. out-scroll the infinite pagination).

The response body size for a data request is now ~512 kB. Before this
change, it was resting around ~1 MB which was unacceptable.