feat(routes/show): add designer select for curators
This patch adds a designer select to the show page to allow curators to
select a designer for a show. Note that this assumes that there is only
one collection per show. As of now, that assumption should always be
true. There are a few notable exceptions that should be manually updated
(e.g. fashion shows that present multiple brands) but this works for
now.

Closes: NC-750