feat: add shows page specific to a single brand
This patch adds a shows page that only displays the shows for a single
brand. I should refactor this to reuse the same components as the main
shows list. To do so, I probably want to update the `<Carousel>` to not
require an explicit width.

Eventually, this page should be replaced with a dedicated brand page
that shows everything about a brand (not just shows).

This page was inspired by the Vogue "designer" page.

Ref: https://www.vogue.com/fashion-shows/designer/off-white