feat: add filters bar component to shows page
This patch refactors the products page to make the filters bar reusable.
I then reuse that filters bar and the filters parsing logic in the shows
page. This is not a perfect adaptation: there is still quite a bit of
logic hard-coded to products in the `<Filters>` definition, but it
mostly works and leads to very little code duplication (and UX
consistency).

This patch also removes the max width from the shows page in favor of
dynamically controlled zoom settings (to increase/decrease the number of
shows per row).

Closes: NC-664