refactor: create reusable `useLoadFetcher` hook
This patch creates a reusable `useLoadFetcher` hook for loading data
given an endpoint string and then re-loading that data whenever the
given endpoint changes. This pattern is already used twice in the app
and will probably be used many times (e.g. when filtering results based
on a user provided query string that can change on every keystroke).