fix: hide articles and reviews from show filters
This patch hides the articles, reviews, and integer ID columns from the
show filters drop-down. No need to surface those to a user.

Perhaps in the future it may become useful to filter for e.g. shows that
have more than XX number of reviews, but for now, that doesn't matter
too much and would be more difficult than it is worth to implement.