feat(show): redesign show page to focus on video
This patch introduces a pretty major redesign to the show page. Instead
of showing the video at the top, in the header, I now show the video as
the main content with the looks scrolling horizontally underneath.
Eventually, I may have the looks auto-scroll based on the video playback
position and vice versa.

This patch also adds resizable panes to this show page: The left pane
(the much larger pane) contains the video (if applicable) and look
images. The right (aside) pane contains all the show information,
reviews, scores, etc.

This change was inspired by Linear's project view (where the main pane
is the project ticket's and the left aside is the project information).

This redesign allows you to fully immerse yourself in the show
experience (i.e. by watching the video) while scrolling through looks
and reading reviews all at the same time in a single view.