feat(show): make article headers sticky
This patch makes the critic review article headers sticky so that when
users scroll through to read them, they always have context as to who
wrote it and for which publication. Normally, this information is
available in the publication's website header or URL. On my site, this
information is now available persistently in the sticky article header.