feat(products): move product dialog to separate page
This patch moves the product dialog to a separate independent page
(similar to the show page that opens when a user clicks on a show). I
may add a "quick view" preview feature in the future, but having a
separate product page makes more sense for now.

This product page should grow fairly big in the foreseeable future.