fix(filters/variant): filter for multiple colors
This patch updates the variant filtering to work properly when selecting
a variant with multiple colors. It will filter for products that contain
a single variant that has both colors.

This is different behavior than selecting two different colors
separately and filtering for products that have variants with either
color in them.