feat(routes/variants): search variants by color name
This patch updates the variant filter route to get results by searching
for the given string in the color names. This is not a perfect solution
as it only works once a user inputs a full color name (e.g. "B" will
have no results while "Blue" will show a bunch).

Ref: https://www.prisma.io/docs/concepts/components/prisma-client/full-text-search