feat(product): show prices per variant option
This patch adds little price range indicators underneath each size and
color option. This UX mimics the experience that StockX provides its
users. If there is no price for a size / color combination, N/A is
shown. Perhaps we should update it to read "SOLD OUT" or "Unavailable"
or provide some type of "BID" or "Notify me" system.